living-system-prompt/requirements.txt
# ============================================================
# requirements.txt — Living System Prompt
# Install active libraries only: pip install -r requirements.txt
# Toggle libraries in /vault/libraries/ before installing.
# ============================================================

# --- Core ---
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.26.0
scipy>=1.12.0

# --- Market Data (default: active) ---
yfinance>=0.2.38

# --- Options Pricing & Greeks (default: active) ---
py_vollib>=1.0.1

# --- IBKR Connectivity (default: active) ---
ib_insync>=0.9.86

# --- Backtesting (default: active) ---
vectorbt>=0.26.2

# --- Streaming (default: active) ---
websockets>=12.0

# --- ML / Scoring (default: active) ---
scikit-learn>=1.4.0

# --- Visualization (default: active) ---
streamlit>=1.33.0
plotly>=5.20.0

# --- Utilities ---
watchdog>=4.0.0        # vault_watcher.py file sync
requests>=2.31.0
pydantic>=2.0.0

# --- Candidates (install manually when toggling on) ---
# QuantLib-Python>=1.34
# openbb>=4.0.0
# confluent-kafka>=2.3.0
# xgboost>=2.0.0
# lightgbm>=4.3.0
# dash>=2.16.0
